main.lua: change nThreads default to 4
With 1 thread, the majority of training time seems to be spent waiting on disk/image IO. Increasing to n=2 or n=4 fixes this, leading to DataTime values consistently ~0s and so faster training which is more CPU/GPU-bound. This would be true of almost all computers, so it makes little sense to have the pessimal default of 1.